feat: update provider interfaces to use UnifiedModelCatalogProviderContext and improve test stubs
