+++ provider/test-harness.ts
@@ -1,6 +1,23 @@
 import { vi } from "vitest";
 import plugin from "../index.js";
-import type { NanoGptModelCatalogProviderRegistration, NanoGptProviderRegistration } from "./types.js";
+import type { UnifiedModelCatalogEntry, UnifiedModelCatalogProviderContext } from "openclaw/plugin-sdk/provider-model-shared";
+import type { NanoGptProviderRegistration } from "./types.js";
+
+/**
+ * Test-only mirror of the shape returned by `api.registerModelCatalogProvider`.
+ * Kept here instead of `types.ts` to avoid widening the public type surface
+ * with a test-only concern.
+ */
+export interface NanoGptModelCatalogProviderRegistration {
+  provider: string;
+  kinds: readonly string[];
+  liveCatalog?: (
+    ctx: UnifiedModelCatalogProviderContext,
+  ) => Promise<readonly UnifiedModelCatalogEntry[]> | readonly UnifiedModelCatalogEntry[];
+  staticCatalog?: (
+    ctx: UnifiedModelCatalogProviderContext,
+  ) => Promise<readonly UnifiedModelCatalogEntry[]> | readonly UnifiedModelCatalogEntry[];
+}
 
 export function getRegisteredProviderHarness(overrideConfig: Record<string, unknown> = {}) {
   const providers: unknown[] = [];
